pyproject.toml
[tool.poetry]
name = "tg-tui"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

packages = [
	{ include = "tg_tui" },
]
include = ["libtdjson.so"]

[tool.poetry.dependencies]
python = "^3.9"
textual = "^0.1.18"
python-telegram = "^0.18"

[tool.poetry.dev-dependencies]
ipython = "^8.3.0"
flakehell = "^0.9.0"
flake8-isort = "^4.1.1"
flake8-bandit = "^3.0.0"
flake8-variables-names = "^0.0.5"
flake8-builtins = "^1.5.3"
flake8-annotations = "^2.9.0"
flake8-annotations-complexity = "^0.0.7"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
line_length = 100
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 3
skip_glob = "**/migrations/*"

[tool.flakehell]
exclude = [".git", "__pycache__", "migrations"]
max_line_length = 100
format = "grouped"
show_source = true

[tool.flakehell.plugins]
pyflakes = ["+*", "-F403", "-F405"]
pycodestyle = ["+*", "-W503"]
flake8-annotations = ["+*", "-ANN002", "-ANN003", "-ANN101"]
"flake8-*" = ["+*"]
mccabe = ["+*"]

[tool.poetry.scripts]
tui = "tg_tui.tui:main"
